+++ gui/Pages/ConjugateHeatTransferView.py
@@ -47,116 +47,28 @@ from code_saturne.Base.QtCore    import *
 from code_saturne.Base.QtGui     import *
 from code_saturne.Base.QtWidgets import *
 
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 # Application modules import
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 
 from code_saturne.model.Common import LABEL_LENGTH_MAX, GuiParam
 from code_saturne.Base.QtPage import IntValidator, DoubleValidator, RegExpValidator, ComboModel
 from code_saturne.Base.QtPage import from_qvariant, to_text_string
 from code_saturne.Pages.ConjugateHeatTransferForm import Ui_ConjugateHeatTransferForm
 from code_saturne.model.ConjugateHeatTransferModel import ConjugateHeatTransferModel
+from code_saturne.model.LocalizationModel import LocalizationModel
+from code_saturne.model.Boundary import Boundary
 
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 # log config
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 
 logging.basicConfig()
 log = logging.getLogger("ConjugateHeatTransferView")
 log.setLevel(GuiParam.DEBUG)
 
-#-------------------------------------------------------------------------------
-# QLineEdit delegate for validation of Syrthes verbosity or visualization
-#-------------------------------------------------------------------------------
 
-class SyrthesVerbosityDelegate(QItemDelegate):
-    def __init__(self, parent = None):
-        super(SyrthesVerbosityDelegate, self).__init__(parent)
-        self.parent = parent
-
-
-    def createEditor(self, parent, option, index):
-        editor = QLineEdit(parent)
-        validator = IntValidator(editor, min=0)
-        editor.setValidator(validator)
-        editor.installEventFilter(self)
-        return editor
-
-
-    def setEditorData(self, editor, index):
-        editor.setAutoFillBackground(True)
-        value = from_qvariant(index.model().data(index, Qt.DisplayRole), to_text_string)
-        editor.setText(value)
-
-
-    def setModelData(self, editor, model, index):
-        if editor.validator().state == QValidator.Acceptable:
-            value = from_qvariant(editor.text(), int)
-            model.setData(index, value, Qt.DisplayRole)
-
-#-------------------------------------------------------------------------------
-# QComboBox delegate for Axis Projection in Conjugate Heat Transfer table
-#-------------------------------------------------------------------------------
-
-class ProjectionAxisDelegate(QItemDelegate):
-    """
-    Use of a combo box in the table.
-    """
-    def __init__(self, parent = None):
-        super(ProjectionAxisDelegate, self).__init__(parent)
-        self.parent = parent
-
-
-    def createEditor(self, parent, option, index):
-        editor = QComboBox(parent)
-        editor.addItem("off")
-        editor.addItem("X")
-        editor.addItem("Y")
-        editor.addItem("Z")
-        editor.installEventFilter(self)
-        return editor
-
-
-    def setEditorData(self, comboBox, index):
-        row = index.row()
-        col = index.column()
-        string = index.model().dataSyrthes[row][col]
-        comboBox.setEditText(string)
-
-
-    def setModelData(self, comboBox, model, index):
-        value = comboBox.currentText()
-        model.setData(index, value, Qt.DisplayRole)
-
-#-------------------------------------------------------------------------------
-# QLineEdit delegate for location
-#-------------------------------------------------------------------------------
-
-class SelectionCriteriaDelegate(QItemDelegate):
-    def __init__(self, parent, mdl):
-        super(SelectionCriteriaDelegate, self).__init__(parent)
-        self.parent = parent
-        self.__model = mdl
-
-
-    def createEditor(self, parent, option, index):
-        editor = QLineEdit(parent)
-        return editor
-
-
-    def setEditorData(self, editor, index):
-        editor.setAutoFillBackground(True)
-        self.value = from_qvariant(index.model().data(index, Qt.DisplayRole), to_text_string)
-        editor.setText(self.value)
-
-
-    def setModelData(self, editor, model, index):
-        value = editor.text()
-
-        if str(value) != "" :
-            model.setData(index, value, Qt.DisplayRole)
-
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 # StandarItemModel class
 #-------------------------------------------------------------------------------
 
@@ -166,17 +78,13 @@ class StandardItemModelSyrthes(QStandardItemModel):
         """
         """
         QStandardItemModel.__init__(self)
-        self.setColumnCount(5)
+        self.setColumnCount(1)
         self.headers = [self.tr("Instance name"),
-                        self.tr("Verbosity"),
-                        self.tr("Visualization"),
-                        self.tr("Projection Axis"),
-                        self.tr("Selection criteria")]
+                        self.tr("Boundary zones")
+                        ]
         self.tooltip = [self.tr("Name of coupled instance"),
-                        self.tr("Verbosity level"),
-                        self.tr("Visualization output level (0 for none)"),
-                        self.tr("Projection axis to match 2D Solid domain"),
-                        self.tr("Selection criteria for coupled boundary faces")]
+                        self.tr("Boundary zones where the selected instance applies")
+                        ]
         self.setColumnCount(len(self.headers))
         self.dataSyrthes = []
         self.__model = model
@@ -197,7 +105,7 @@ class StandardItemModelSyrthes(QStandardItemModel):
     def flags(self, index):
         if not index.isValid():
             return Qt.ItemIsEnabled
-        return Qt.ItemIsEnabled | Qt.ItemIsSelectable | Qt.ItemIsEditable
+        return Qt.ItemIsEnabled
 
 
     def headerData(self, section, orientation, role):
@@ -211,34 +119,19 @@ class StandardItemModelSyrthes(QStandardItemModel):
             return
 
         row = index.row()
-        if index.column() in (0, 3, 4):
-            self.dataSyrthes[row][index.column()] = str(from_qvariant(value, to_text_string))
-        else:
-            self.dataSyrthes[row][index.column()] = from_qvariant(value, int)
-
-        num = row + 1
-        self.__model.setSyrthesInstanceName(num, self.dataSyrthes[row][0])
-        self.__model.setSyrthesVerbosity(num, self.dataSyrthes[row][1])
-        self.__model.setSyrthesVisualization(num, self.dataSyrthes[row][2])
-        self.__model.setSyrthesProjectionAxis(num, self.dataSyrthes[row][3])
-        self.__model.setSelectionCriteria(num, self.dataSyrthes[row][4])
-
+        self.dataSyrthes[row][index.column()] = str(from_qvariant(value, to_text_string))
         id1 = self.index(0, 0)
         id2 = self.index(self.rowCount(), 0)
         self.dataChanged.emit(id1, id2)
         return True
 
-
-    def addItem(self, syrthes_name,
-                verbosity, visualization, proj_axis, location):
+    def addItem(self, row):
         """
         Add a row in the table.
         """
-        self.dataSyrthes.append([syrthes_name,
-                                 verbosity, visualization, proj_axis, location])
+        self.dataSyrthes.append(row)
         row = self.rowCount()
         self.setRowCount(row+1)
-
 
     def deleteRow(self, row):
         """
@@ -246,16 +139,18 @@ class StandardItemModelSyrthes(QStandardItemModel):
         """
         del self.dataSyrthes[row]
         row = self.rowCount()
-        self.setRowCount(row-1)
+        self.setRowCount(row - 1)
 
-#-------------------------------------------------------------------------------
+
+# -------------------------------------------------------------------------------
 # Main class
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 
 class ConjugateHeatTransferView(QWidget, Ui_ConjugateHeatTransferForm):
     """
     """
-    def __init__(self, parent, case):
+
+    def __init__(self, parent=None):
         """
         Constructor
         """
@@ -263,6 +158,8 @@ class ConjugateHeatTransferView(QWidget, Ui_ConjugateHeatTransferForm):
 
         Ui_ConjugateHeatTransferForm.__init__(self)
         self.setupUi(self)
+
+    def setup(self, case):
 
         self.case = case
 
@@ -277,73 +174,68 @@ class ConjugateHeatTransferView(QWidget, Ui_ConjugateHeatTransferForm):
         if QT_API == "PYQT4":
             self.tableViewSyrthes.verticalHeader().setResizeMode(QHeaderView.ResizeToContents)
             self.tableViewSyrthes.horizontalHeader().setResizeMode(QHeaderView.ResizeToContents)
-            self.tableViewSyrthes.horizontalHeader().setResizeMode(4, QHeaderView.Stretch)
+            self.tableViewSyrthes.horizontalHeader().setResizeMode(1, QHeaderView.Stretch)
         elif QT_API == "PYQT5":
             self.tableViewSyrthes.verticalHeader().setSectionResizeMode(QHeaderView.ResizeToContents)
             self.tableViewSyrthes.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeToContents)
-            self.tableViewSyrthes.horizontalHeader().setSectionResizeMode(4, QHeaderView.Stretch)
+            self.tableViewSyrthes.horizontalHeader().setSectionResizeMode(1, QHeaderView.Stretch)
 
-        delegateSyrthesVerbosity = SyrthesVerbosityDelegate(self.tableViewSyrthes)
-        self.tableViewSyrthes.setItemDelegateForColumn(1, delegateSyrthesVerbosity)
-        self.tableViewSyrthes.setItemDelegateForColumn(2, delegateSyrthesVerbosity)
-        delegateProjectionAxis = ProjectionAxisDelegate(self.tableViewSyrthes)
-        self.tableViewSyrthes.setItemDelegateForColumn(3, delegateProjectionAxis)
-        delegateSelectionCriteria = SelectionCriteriaDelegate(self.tableViewSyrthes, self.__model)
-        self.tableViewSyrthes.setItemDelegateForColumn(4, delegateSelectionCriteria)
+        self.modelProjectionAxis = ComboModel(self.comboBoxProjectionAxis, 1, 1)
+        self.modelProjectionAxis.addItem("off")
+        self.modelProjectionAxis.addItem("x")
+        self.modelProjectionAxis.addItem("y")
+        self.modelProjectionAxis.addItem("z")
 
-        # Connections
-        self.pushButtonAdd.clicked.connect(self.slotAddSyrthes)
-        self.pushButtonDelete.clicked.connect(self.slotDeleteSyrthes)
+        # connections
+        self.comboBoxProjectionAxis.currentTextChanged[str].connect(self.slotProjectionAxis)
+        self.lineEditVerbosity.textChanged[str].connect(self.slotVerbosity)
+        self.lineEditVisualization.textChanged[str].connect(self.slotVisualization)
+        self.lineEditVerbosity.setValidator(IntValidator(self.lineEditVerbosity))
+        self.lineEditVisualization.setValidator(IntValidator(self.lineEditVisualization))
 
-        # Insert list of Syrthes couplings for view
-        for c in self.__model.getSyrthesCouplingList():
-            [syrthes_name, verbosity, visualization, proj_axis, location] = c
-            self.modelSyrthes.addItem(syrthes_name,
-                                      verbosity, visualization, proj_axis, location)
-
-        if len(self.__model.getSyrthesCouplingList()) < 2:
-            self.tableViewSyrthes.hideColumn(0)
+        self.initializeParameters()
 
         self.case.undoStartGlobal()
 
-
-    @pyqtSlot()
-    def slotAddSyrthes(self):
-        """
-        Set in view label and variables to see on profile
-        """
-        syrthes_name  = self.__model.defaultValues()['syrthes_name']
-        verbosity     = self.__model.defaultValues()['verbosity']
-        visualization = self.__model.defaultValues()['visualization']
-        proj_axis     = self.__model.defaultValues()['projection_axis']
-        location      = self.__model.defaultValues()['selection_criteria']
-        num = self.__model.addSyrthesCoupling(syrthes_name,
-                                              verbosity, visualization,
-                                              proj_axis, location)
-        self.modelSyrthes.addItem(syrthes_name, verbosity, visualization,
-                                  proj_axis, location)
-        if len(self.__model.getSyrthesCouplingList()) > 1:
-            self.tableViewSyrthes.showColumn(0)
-
-
-    @pyqtSlot()
-    def slotDeleteSyrthes(self):
-        """
-        Delete the profile from the list (one by one).
-        """
-        row = self.tableViewSyrthes.currentIndex().row()
-        log.debug("slotDeleteSyrthes -> %s" % (row,))
-        if row == -1:
-            title = self.tr("Warning")
-            msg   = self.tr("You must select an existing coupling")
-            QMessageBox.information(self, title, msg)
+    def initializeParameters(self):
+        projection_axis = self.__model.getSyrthesProjectionAxis()
+        if projection_axis not in ["", None]:
+            self.comboBoxProjectionAxis.setCurrentText(projection_axis)
         else:
-            self.modelSyrthes.deleteRow(row)
-            self.__model.deleteSyrthesCoupling(row+1)
-            if len(self.__model.getSyrthesCouplingList()) < 2:
-                self.tableViewSyrthes.hideColumn(0)
+            self.comboBoxProjectionAxis.setCurrentText("off")
+            self.slotProjectionAxis("off")
+
+        verbosity = self.__model.getSyrthesVerbosity()
+        if verbosity not in ["", None]:
+            self.lineEditVerbosity.setText(verbosity)
+        else:
+            self.lineEditVerbosity.setText("0")
+
+        visualization = self.__model.getSyrthesVisualization()
+        if visualization not in ["", None]:
+            self.lineEditVisualization.setText(visualization)
+        else:
+            self.lineEditVisualization.setText("1")
 
 
-#-------------------------------------------------------------------------------
+        for syrthes_name, boundary_labels in self.__model.getSyrthesCouplingList():
+            self.modelSyrthes.addItem([syrthes_name, ", ".join(boundary_labels)])
+
+    @pyqtSlot(str)
+    def slotProjectionAxis(self, value):
+        self.__model.setSyrthesProjectionAxis(value)
+        return
+
+    @pyqtSlot(str)
+    def slotVerbosity(self, value):
+        self.__model.setSyrthesVerbosity(value)
+        pass
+
+    @pyqtSlot(str)
+    def slotVisualization(self, value):
+        self.__model.setSyrthesVisualization(value)
+        pass
+
+# -------------------------------------------------------------------------------
 # End
 #-------------------------------------------------------------------------------